Adds -Wextra to configure to detect possible bugs, such as initialized variables. (It also adds -Wno-unused-parameter, since it unused parameters do not cause bugs, and to silence the warning we would have to write (void)(parameter);, for every that is not used, which detracts from the readability of the code.)
Once -Wextra was added, the compiler mostly found comparisons
between signed and unsigned numbers.



git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2642 a4d8336d-3463-0410-8bba-c098c45d37a8

+++ allpairs/src/allpairs_master.c
@@ -164,7 +164,7 @@ char * text_list_string( struct text_list *t, int a, int b )
 		const char *str = text_list_get(t,i);
 		if(!str) break;
 		str = string_basename(str);
-		while((strlen(str) + buffer_pos + 3)>= buffer_size) {
+		while( (int) (strlen(str) + buffer_pos + 3) >= buffer_size) {
 			buffer_size *= 2;
 			buffer = realloc(buffer,buffer_size);
 		}
